backend/requirements.txt
fastapi
uvicorn
pypdf2
python-multipart
requests
google-genai
chromadb
python-dotenv